requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
google-generativeai==0.1.0
mangum==0.15.0
youtube_transcript_api==0.4.3  # if needed by your project